C# DB Advanced - Entity Framework -05.Entity-Relations

+++ C#/DB/Entity Framework Core/05.Entity-Relations/03.FootballBetting/P03_FootballBetting.Data.Models/Game.cs
@@ -30,7 +30,7 @@
 
         public double AwayTeamBetRate { get; set; }
 
-        public int DrawBetRate { get; set; }
+        public double DrawBetRate { get; set; }
 
         public string Result { get; set; }
